refactor: move retry job on failure tests
These tests verify that the job behavior of the listener works as
expected for any job. When FAILED with retries, the job can be retried
by another worker.
