+++ zeebe/engine/src/test/java/io/camunda/zeebe/engine/processing/bpmn/activity/listeners/task/TaskListenerTest.java
@@ -76,90 +76,6 @@ public class TaskListenerTest {
   }
 
   @Test
-  public void shouldRetryAssigningListenerWhenListenerJobFailedOnTaskAssignAfterCreation() {
-    verifyListenerIsRetriedWhenListenerJobFailed(
-        ZeebeTaskListenerEventType.assigning,
-        userTask -> userTask.zeebeAssignee("gandalf"),
-        userTaskClient -> {},
-        UserTaskIntent.ASSIGNED);
-  }
-
-  @Test
-  public void shouldRetryAssigningListenerWhenListenerJobFailedOnTaskAssign() {
-    verifyListenerIsRetriedWhenListenerJobFailed(
-        ZeebeTaskListenerEventType.assigning,
-        UnaryOperator.identity(),
-        userTaskClient -> userTaskClient.withAssignee("bilbo").assign(),
-        UserTaskIntent.ASSIGNED);
-  }
-
-  @Test
-  public void shouldRetryAssigningListenerWhenListenerJobFailedOnTaskClaim() {
-    verifyListenerIsRetriedWhenListenerJobFailed(
-        ZeebeTaskListenerEventType.assigning,
-        UnaryOperator.identity(),
-        userTaskClient -> userTaskClient.withAssignee("bilbo").claim(),
-        UserTaskIntent.ASSIGNED);
-  }
-
-  @Test
-  public void shouldRetryUpdatingListenerWhenListenerJobFailedOnTaskUpdate() {
-    verifyListenerIsRetriedWhenListenerJobFailed(
-        ZeebeTaskListenerEventType.updating,
-        UnaryOperator.identity(),
-        UserTaskClient::update,
-        UserTaskIntent.UPDATED);
-  }
-
-  @Test
-  public void shouldRetryCompletingListenerWhenListenerJobFailedOnTaskComplete() {
-    verifyListenerIsRetriedWhenListenerJobFailed(
-        ZeebeTaskListenerEventType.completing,
-        UnaryOperator.identity(),
-        UserTaskClient::complete,
-        UserTaskIntent.COMPLETED);
-  }
-
-  private void verifyListenerIsRetriedWhenListenerJobFailed(
-      final ZeebeTaskListenerEventType eventType,
-      final UnaryOperator<UserTaskBuilder> userTaskBuilder,
-      final Consumer<UserTaskClient> userTaskAction,
-      final UserTaskIntent terminalActionIntent) {
-    // given
-    final long processInstanceKey =
-        helper.createProcessInstance(
-            helper.createProcessWithZeebeUserTask(
-                t ->
-                    userTaskBuilder
-                        .apply(t)
-                        .zeebeTaskListener(l -> l.eventType(eventType).type(listenerType))
-                        .zeebeTaskListener(l -> l.eventType(eventType).type(listenerType + "_2"))));
-
-    // when: performing the user task action
-    userTaskAction.accept(ENGINE.userTask().ofInstance(processInstanceKey));
-
-    // when: fail listener job with retries
-    ENGINE.job().ofInstance(processInstanceKey).withType(listenerType).withRetries(1).fail();
-    // complete failed and remaining listeners job
-    helper.completeJobs(processInstanceKey, listenerType, listenerType + "_2");
-
-    // then: assert the listener job was completed after the failure
-    assertThat(records().limit(r -> r.getIntent() == terminalActionIntent))
-        .extracting(Record::getValueType, Record::getIntent)
-        .containsSubsequence(
-            tuple(ValueType.JOB, JobIntent.CREATED),
-            tuple(ValueType.JOB, JobIntent.FAILED),
-            tuple(ValueType.JOB, JobIntent.COMPLETE),
-            tuple(ValueType.JOB, JobIntent.COMPLETED),
-            tuple(ValueType.USER_TASK, UserTaskIntent.COMPLETE_TASK_LISTENER),
-            tuple(ValueType.JOB, JobIntent.CREATED),
-            tuple(ValueType.JOB, JobIntent.COMPLETE),
-            tuple(ValueType.JOB, JobIntent.COMPLETED),
-            tuple(ValueType.USER_TASK, UserTaskIntent.COMPLETE_TASK_LISTENER),
-            tuple(ValueType.USER_TASK, terminalActionIntent));
-  }
-
-  @Test
   public void
       shouldCreateJobNoRetriesIncidentForAssigningListenerAndContinueAfterResolutionOnTaskAssignAfterCreation() {
     verifyIncidentCreationOnListenerJobWithoutRetriesAndResolution(
